+++ config/max-lines-baseline.txt
@@ -167,7 +167,6 @@ inline src/main/window/attach-main-window-services.ts
 inline src/main/window/createMainWindow.test.ts
 inline src/main/window/createMainWindow.ts
 inline src/main/workspace-space-analysis.ts
-inline src/preload/api-types.ts
 inline src/preload/index.ts
 inline src/relay/agent-hook-server.ts
 inline src/relay/dispatcher.ts
