+++ config/tsconfig.tc.web.json
@@ -5,6 +5,7 @@
     "../src/renderer/src/**/*",
     "../src/renderer/src/**/*.tsx",
     "../src/preload/api-types.ts",
+    "../src/preload/api/**/*",
     "../src/shared/**/*",
     "../src/main/gitlab/mappers.ts",
     "../src/main/ipc/worktree-branch-name.ts",
