+++ src/preload/api/orca-profile-api.ts
@@ -1,0 +1,59 @@
+import type {
+  ConnectCurrentOrcaProfileResult,
+  CreateCloudLinkedOrcaProfileArgs,
+  CreateCloudLinkedOrcaProfileResult,
+  CreateLocalOrcaProfileArgs,
+  CreateLocalOrcaProfileResult,
+  FindOrcaProfileProjectsByPathArgs,
+  FindOrcaProfileProjectsByPathResult,
+  OrcaProfileAuthStatus,
+  OrcaProfileListResult,
+  OrcaProfileOrgInviteRevokeArgs,
+  OrcaProfileOrgMemberChangeRoleArgs,
+  OrcaProfileOrgMemberInviteArgs,
+  OrcaProfileOrgMemberMutationResult,
+  OrcaProfileOrgMemberRemoveArgs,
+  OrcaProfileOrgMembersListArgs,
+  OrcaProfileOrgMembersListResult,
+  RefreshCurrentOrcaProfileAuthResult,
+  SelectOrcaProfileOrgArgs,
+  SelectOrcaProfileOrgResult,
+  SignOutCurrentOrcaProfileResult,
+  SwitchOrcaProfileArgs,
+  SwitchOrcaProfileResult,
+  TransferOrcaProfileProjectArgs,
+  TransferOrcaProfileProjectResult
+} from '../../shared/orca-profiles'
+
+export type OrcaProfileApi = {
+  list: () => Promise<OrcaProfileListResult>
+  authStatus: () => Promise<OrcaProfileAuthStatus>
+  createLocal: (args?: CreateLocalOrcaProfileArgs) => Promise<CreateLocalOrcaProfileResult>
+  createCloudLinked: (
+    args?: CreateCloudLinkedOrcaProfileArgs
+  ) => Promise<CreateCloudLinkedOrcaProfileResult>
+  switchProfile: (args: SwitchOrcaProfileArgs) => Promise<SwitchOrcaProfileResult>
+  transferProject: (
+    args: TransferOrcaProfileProjectArgs
+  ) => Promise<TransferOrcaProfileProjectResult>
+  findProjectProfiles: (
+    args: FindOrcaProfileProjectsByPathArgs
+  ) => Promise<FindOrcaProfileProjectsByPathResult>
+  connectCurrent: () => Promise<ConnectCurrentOrcaProfileResult>
+  refreshAuth: () => Promise<RefreshCurrentOrcaProfileAuthResult>
+  signOutCurrent: () => Promise<SignOutCurrentOrcaProfileResult>
+  selectOrg: (args: SelectOrcaProfileOrgArgs) => Promise<SelectOrcaProfileOrgResult>
+  orgMembersList: (args: OrcaProfileOrgMembersListArgs) => Promise<OrcaProfileOrgMembersListResult>
+  orgMemberInvite: (
+    args: OrcaProfileOrgMemberInviteArgs
+  ) => Promise<OrcaProfileOrgMemberMutationResult>
+  orgInviteRevoke: (
+    args: OrcaProfileOrgInviteRevokeArgs
+  ) => Promise<OrcaProfileOrgMemberMutationResult>
+  orgMemberChangeRole: (
+    args: OrcaProfileOrgMemberChangeRoleArgs
+  ) => Promise<OrcaProfileOrgMemberMutationResult>
+  orgMemberRemove: (
+    args: OrcaProfileOrgMemberRemoveArgs
+  ) => Promise<OrcaProfileOrgMemberMutationResult>
+}
