+++ src/preload/api/ui-window-api.ts
@@ -1,0 +1,51 @@
+import type { ReadClipboardTextOptions } from '../../shared/clipboard-text'
+import type { NativeFileDropPayload } from '../../shared/native-file-drop'
+import type {
+  RichMarkdownContextMenuCommandPayload,
+  RichMarkdownContextMenuTableTarget
+} from '../../shared/rich-markdown-context-menu'
+
+export type UiWindowApi = {
+  readClipboardText: (options?: ReadClipboardTextOptions) => Promise<string>
+  readSelectionClipboardText: (options?: ReadClipboardTextOptions) => Promise<string>
+  saveClipboardImageAsTempFile: (args?: {
+    connectionId?: string | null
+    runtimeEnvironmentId?: string | null
+  }) => Promise<string | null>
+  writeClipboardText: (text: string) => Promise<void>
+  writeTerminalClipboardText: (text: string) => Promise<void>
+  writeSelectionClipboardText: (text: string) => Promise<void>
+  writeClipboardImage: (dataUrl: string) => Promise<void>
+  performNativePaste: (options?: { mode?: 'paste' | 'paste-and-match-style' }) => void
+  performNativeSelectionAction: (action: 'copy' | 'select-all') => void
+  writeClipboardFile: (
+    args:
+      | {
+          filePath: string
+          connectionId?: string | null
+        }
+      | string
+  ) => Promise<{ ok: boolean; reason?: string }>
+  onFileDrop: (callback: (data: NativeFileDropPayload) => void) => () => void
+  getZoomLevel: () => number
+  setZoomLevel: (level: number) => void
+  syncTrafficLights: (zoomFactor: number) => void
+  setMarkdownEditorFocused: (focused: boolean) => void
+  setRichMarkdownContextMenuTarget: (target: RichMarkdownContextMenuTableTarget | null) => void
+  setTerminalInputFocused: (focused: boolean) => void
+  setFloatingFocus: (state: { panelFocused: boolean; terminalFocused: boolean }) => void
+  setShortcutRecorderFocused: (focused: boolean) => void
+  onRichMarkdownContextCommand: (
+    callback: (payload: RichMarkdownContextMenuCommandPayload) => void
+  ) => () => void
+  onFullscreenChanged: (callback: (isFullScreen: boolean) => void) => () => void
+  minimize: () => void
+  maximize: () => void
+  isMaximized: () => Promise<boolean>
+  onMaximizeChanged: (callback: (isMaximized: boolean) => void) => () => void
+  requestClose: () => void
+  popupMenu: () => void
+  onWindowCloseRequested: (callback: (data: { isQuitting: boolean }) => void) => () => void
+  confirmWindowClose: () => void
+  notifyWindowRevealed: () => void
+}
