+++ src/preload/api/updater-api.ts
@@ -1,0 +1,25 @@
+import type {
+  LinuxPackageInstallInstructions,
+  ReleaseBuildListResult,
+  UpdateCheckOptions,
+  UpdateStatus
+} from '../../shared/types'
+import type { ReleaseChannel } from '../../shared/release-channel'
+
+export type UpdaterApi = {
+  getVersion: () => Promise<string>
+  getStatus: () => Promise<UpdateStatus>
+  check: (options?: UpdateCheckOptions) => Promise<void>
+  download: () => Promise<void>
+  quitAndInstall: () => Promise<void>
+  dismissNudge: () => Promise<void>
+  dismissAvailableUpdate: () => Promise<void>
+  /** Desktop-only. Rejects unless the current status carries `linux-package-install` recovery. */
+  getLinuxPackageInstallInstructions: () => Promise<LinuxPackageInstallInstructions>
+  /** Desktop-only. Reveals the revalidated cached package in the native file manager. */
+  showLinuxPackage: () => Promise<void>
+  listBuilds: (channel: ReleaseChannel) => Promise<ReleaseBuildListResult>
+
+  onStatus: (callback: (status: UpdateStatus) => void) => () => void
+  onClearDismissal: (callback: () => void) => () => void
+}
