+++ src/preload/api/workspace-cleanup-api.ts
@@ -1,0 +1,30 @@
+import type {
+  WorkspaceCleanupDismissArgs,
+  WorkspaceCleanupLocalProcessArgs,
+  WorkspaceCleanupLocalProcessResult,
+  WorkspaceCleanupScanArgs,
+  WorkspaceCleanupScanProgress,
+  WorkspaceCleanupScanResult
+} from '../../shared/workspace-cleanup'
+import type {
+  WorkspaceSpaceAnalyzeResult,
+  WorkspaceSpaceScanProgress
+} from '../../shared/workspace-space-types'
+
+export type WorkspaceCleanupApi = {
+  scan: (
+    args?: WorkspaceCleanupScanArgs,
+    onProgress?: (progress: WorkspaceCleanupScanProgress) => void
+  ) => Promise<WorkspaceCleanupScanResult>
+  dismiss: (args: WorkspaceCleanupDismissArgs) => Promise<void>
+  clearDismissals: () => Promise<void>
+  hasKillableLocalProcesses: (
+    args: WorkspaceCleanupLocalProcessArgs
+  ) => Promise<WorkspaceCleanupLocalProcessResult>
+}
+
+export type WorkspaceSpaceApi = {
+  analyze: () => Promise<WorkspaceSpaceAnalyzeResult>
+  cancel: () => Promise<boolean>
+  onProgress: (callback: (progress: WorkspaceSpaceScanProgress) => void) => () => void
+}
